API review followups (#50181)
* RazorComponentResult: change namespace, make executor internal, merge test classes

* RazorComponentResult: Nullability and trim annotations

* RazorComponentResult: IStatusCodeHttpResult and IContentTypeHttpResult

* Cleanup

* Further clean up annotations

* Remove HtmlRootComponent.ComponentId as per API review

* Rename SupplyParameterFromFormAttribute.Handler to FormName

* API review: seal

* Clarify RenderModeAttribute inheritance. Fixes #49848

* Rename valueFactory -> initialValueFactory

* Remove unnecessary sequence params

* Make [StreamRendering] default to true
