+++ src/Components/Components/src/RenderTree/RenderTreeFrameArrayBuilder.cs
@@ -135,7 +135,7 @@ internal sealed class RenderTreeFrameArrayBuilder : ArrayBuilder<RenderTreeFrame
         };
     }
 
-    public void AppendComponentRenderMode(int sequence, IComponentRenderMode renderMode)
+    public void AppendComponentRenderMode(IComponentRenderMode renderMode)
     {
         if (_itemsInUse == _items.Length)
         {
@@ -144,13 +144,13 @@ internal sealed class RenderTreeFrameArrayBuilder : ArrayBuilder<RenderTreeFrame
 
         _items[_itemsInUse++] = new RenderTreeFrame
         {
-            SequenceField = sequence,
+            SequenceField = 0, // We're only interested in one of these, so it's not useful to optimize diffing over multiple
             FrameTypeField = RenderTreeFrameType.ComponentRenderMode,
             ComponentRenderModeField = renderMode,
         };
     }
 
-    public void AppendNamedEvent(int sequence, string eventType, string assignedName)
+    public void AppendNamedEvent(string eventType, string assignedName)
     {
         if (_itemsInUse == _items.Length)
         {
@@ -159,7 +159,7 @@ internal sealed class RenderTreeFrameArrayBuilder : ArrayBuilder<RenderTreeFrame
 
         _items[_itemsInUse++] = new RenderTreeFrame
         {
-            SequenceField = sequence,
+            SequenceField = 0, // We're only interested in one of these per eventType, so it's not useful to optimize diffing over multiple
             FrameTypeField = RenderTreeFrameType.NamedEvent,
             NamedEventTypeField = eventType,
             NamedEventAssignedNameField = assignedName,
