+++ src/Components/Components/src/StreamRenderingAttribute.cs
@@ -18,8 +18,8 @@ public class StreamRenderingAttribute : Attribute
     /// <summary>
     /// Constructs an instance of <see cref="StreamRenderingAttribute"/>
     /// </summary>
-    /// <param name="enabled">A flag to indicate whether this component and its descendants should stream their rendering.</param>
-    public StreamRenderingAttribute(bool enabled)
+    /// <param name="enabled">A flag to indicate whether this component and its descendants should stream their rendering. The default value is true.</param>
+    public StreamRenderingAttribute(bool enabled = true)
     {
         Enabled = enabled;
     }
