+++ src/Components/Components/test/ComponentFactoryTest.cs
@@ -152,6 +152,29 @@ public class ComponentFactoryTest
     }
 
     [Fact]
+    public void InstantiateComponent_WithDerivedRenderModeOnDerivedComponent_CausesAmbiguousMatchException()
+    {
+        // We could allow derived components to override the rendermode, but:
+        // [1] It's unclear how that would be legitimate. If the base specifies a rendermode, it's saying
+        //     it only works in that mode. It wouldn't be safe for a derived type to change that.
+        // [2] If we did want to implement this, we'd need to implement our own inheritance chain walking
+        //     to make sure we find the rendermode from the *closest* ancestor type. GetCustomAttributes
+        //     on its own isn't documented to return the results in any specific order.
+        // Since issue [1] makes it unclear we'd want to support this, for now we don't.
+
+        // Arrange
+        var resolvedComponent = new ComponentWithInjectProperties();
+        var componentType = typeof(DerivedComponentWithRenderMode);
+        var renderer = new RendererWithResolveComponentForRenderMode(resolvedComponent);
+        var componentActivator = new DefaultComponentActivator();
+        var factory = new ComponentFactory(componentActivator, renderer);
+
+        // Act/Assert
+        Assert.Throws<AmbiguousMatchException>(
+            () => factory.InstantiateComponent(GetServiceProvider(), componentType, null, 1234));
+    }
+
+    [Fact]
     public void InstantiateComponent_WithRenderModeOnCallSite_UsesRenderModeResolver()
     {
         // Arrange
@@ -290,6 +313,16 @@ public class ComponentFactoryTest
     }
 
     private class TestRenderMode : IComponentRenderMode { }
+    private class DerivedComponentRenderMode : IComponentRenderMode { }
+
+    [DerivedComponentRenderMode]
+    private class DerivedComponentWithRenderMode : ComponentWithRenderMode
+    {
+        class DerivedComponentRenderModeAttribute : RenderModeAttribute
+        {
+            public override IComponentRenderMode Mode => new DerivedComponentRenderMode();
+        }
+    }
 
     [OwnRenderMode]
     private class ComponentWithRenderMode : IComponent
