+++ src/Components/Endpoints/src/Results/RazorComponentResult.cs
@@ -1,18 +1,19 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
+using System.Collections.ObjectModel;
 using System.Diagnostics.CodeAnalysis;
-using Microsoft.AspNetCore.Http;
-using Microsoft.Extensions.DependencyInjection;
+using Microsoft.AspNetCore.Components;
+using Microsoft.AspNetCore.Components.Endpoints;
 using Microsoft.Extensions.Internal;
 using static Microsoft.AspNetCore.Internal.LinkerFlags;
 
-namespace Microsoft.AspNetCore.Components.Endpoints;
+namespace Microsoft.AspNetCore.Http.HttpResults;
 
 /// <summary>
 /// An <see cref="IResult"/> that renders a Razor Component.
 /// </summary>
-public class RazorComponentResult : IResult
+public class RazorComponentResult : IResult, IStatusCodeHttpResult, IContentTypeHttpResult
 {
     private static readonly IReadOnlyDictionary<string, object?> EmptyParameters
         = new Dictionary<string, object?>().AsReadOnly();
@@ -22,7 +23,7 @@ public class RazorComponentResult : IResult
     /// </summary>
     /// <param name="componentType">The type of the component to render. This must implement <see cref="IComponent"/>.</param>
     public RazorComponentResult([DynamicallyAccessedMembers(Component)] Type componentType)
-        : this(componentType, null)
+        : this(componentType, ReadOnlyDictionary<string, object?>.Empty)
     {
     }
 
@@ -31,7 +32,9 @@ public class RazorComponentResult : IResult
     /// </summary>
     /// <param name="componentType">The type of the component to render. This must implement <see cref="IComponent"/>.</param>
     /// <param name="parameters">Parameters for the component.</param>
-    public RazorComponentResult([DynamicallyAccessedMembers(Component)] Type componentType, object? parameters)
+    public RazorComponentResult(
+        [DynamicallyAccessedMembers(Component)] Type componentType,
+        [DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.All)] object parameters)
         : this(componentType, CoerceParametersObjectToDictionary(parameters))
     {
     }
@@ -41,25 +44,27 @@ public class RazorComponentResult : IResult
     /// </summary>
     /// <param name="componentType">The type of the component to render. This must implement <see cref="IComponent"/>.</param>
     /// <param name="parameters">Parameters for the component.</param>
-    public RazorComponentResult([DynamicallyAccessedMembers(Component)] Type componentType, IReadOnlyDictionary<string, object?>? parameters)
+    public RazorComponentResult(
+        [DynamicallyAccessedMembers(Component)] Type componentType,
+        IReadOnlyDictionary<string, object?> parameters)
     {
+        ArgumentNullException.ThrowIfNull(componentType);
+        ArgumentNullException.ThrowIfNull(parameters);
+
         // Note that the Blazor renderer will validate that componentType implements IComponent and throws a suitable
         // exception if not, so we don't need to duplicate that logic here.
-
-        ArgumentNullException.ThrowIfNull(componentType);
         ComponentType = componentType;
         Parameters = parameters ?? EmptyParameters;
     }
 
-    private static IReadOnlyDictionary<string, object?>? CoerceParametersObjectToDictionary(object? parameters)
+    private static IReadOnlyDictionary<string, object?> CoerceParametersObjectToDictionary(object? parameters)
         => parameters is null
-        ? null
+        ? throw new ArgumentNullException(nameof(parameters))
         : (IReadOnlyDictionary<string, object?>)PropertyHelper.ObjectToDictionary(parameters);
 
     /// <summary>
     /// Gets the component type.
     /// </summary>
-    [DynamicallyAccessedMembers(Component)]
     public Type ComponentType { get; }
 
     /// <summary>
@@ -87,15 +92,10 @@ public class RazorComponentResult : IResult
     public bool PreventStreamingRendering { get; set; }
 
     /// <summary>
-    /// Requests the service of
-    /// <see cref="RazorComponentResultExecutor.ExecuteAsync(HttpContext, RazorComponentResult)" />
-    /// to process itself in the given <paramref name="httpContext" />.
+    /// Processes this result in the given <paramref name="httpContext" />.
     /// </summary>
     /// <param name="httpContext">An <see cref="HttpContext" /> associated with the current request.</param >
     /// <returns >A <see cref="T:System.Threading.Tasks.Task" /> which will complete when execution is completed.</returns >
     public Task ExecuteAsync(HttpContext httpContext)
-    {
-        var executor = httpContext.RequestServices.GetRequiredService<RazorComponentResultExecutor>();
-        return executor.ExecuteAsync(httpContext, this);
-    }
+        => RazorComponentResultExecutor.ExecuteAsync(httpContext, this);
 }
