+++ src/Components/Endpoints/src/Results/RazorComponentResultExecutor.cs
@@ -9,23 +9,15 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.WebUtilities;
 using Microsoft.Extensions.DependencyInjection;
 using static Microsoft.AspNetCore.Internal.LinkerFlags;
+using Microsoft.AspNetCore.Http.HttpResults;
 
 namespace Microsoft.AspNetCore.Components.Endpoints;
 
-/// <summary>
-/// Executes a <see cref="RazorComponentResult"/>.
-/// </summary>
-public class RazorComponentResultExecutor
+internal static class RazorComponentResultExecutor
 {
-    /// <summary>
-    /// The default content-type header value for Razor Components, <c>text/html; charset=utf-8</c>.
-    /// </summary>
-    public static readonly string DefaultContentType = "text/html; charset=utf-8";
+    public const string DefaultContentType = "text/html; charset=utf-8";
 
-    /// <summary>
-    /// Executes a <see cref="RazorComponentResult"/> asynchronously.
-    /// </summary>
-    public virtual Task ExecuteAsync(HttpContext httpContext, RazorComponentResult result)
+    public static Task ExecuteAsync(HttpContext httpContext, RazorComponentResult result)
     {
         ArgumentNullException.ThrowIfNull(httpContext);
 
@@ -44,7 +36,7 @@ public class RazorComponentResultExecutor
             result.PreventStreamingRendering);
     }
 
-    internal static Task RenderComponentToResponse(
+    private static Task RenderComponentToResponse(
         HttpContext httpContext,
         [DynamicallyAccessedMembers(Component)] Type componentType,
         IReadOnlyDictionary<string, object?>? componentParameters,
