+++ src/Components/Endpoints/test/TestComponents/DoubleRenderingStreamingAsyncComponent.razor
@@ -1,4 +1,4 @@
-﻿@attribute [StreamRendering(true)]
+﻿@attribute [StreamRendering]
 @if (loaded)
 {
     <text>Loaded</text>
