+++ src/Components/Endpoints/test/TestComponents/StreamingAsyncLoadingComponent.razor
@@ -1,4 +1,4 @@
-﻿@attribute [StreamRendering(true)]
+﻿@attribute [StreamRendering]
 Loading task status: @LoadingTask.Status
 @code {
     [Parameter] public Task LoadingTask { get; set; }
