+++ src/Components/Endpoints/test/TestComponents/StreamingComponentThatRedirectsAsynchronously.razor
@@ -1,4 +1,4 @@
-﻿@attribute [StreamRendering(true)]
+﻿@attribute [StreamRendering]
 @inject NavigationManager Nav
 Some output
 @code {
