+++ src/Components/Endpoints/test/TestComponents/StreamingComponentThatThrowsAsynchronously.razor
@@ -1,4 +1,4 @@
-﻿@attribute [StreamRendering(true)]
+﻿@attribute [StreamRendering]
 Some output
 @code {
     protected override async Task OnInitializedAsync()
