+++ src/Components/Endpoints/test/TestComponents/StreamingComponentWithChild.razor
@@ -1,4 +1,4 @@
-﻿@attribute [StreamRendering(true)]
+﻿@attribute [StreamRendering]
 [LoadingTask: @LoadingTask.Status]
 <StreamingComponentChild SomeParameterToEnsureRerendering="@(new object())" />
 @code {
