+++ src/Components/Shared/test/AssertFrame.cs
@@ -141,18 +141,18 @@ internal static class AssertFrame
         AssertFrame.Sequence(frame, sequence);
     }
 
-    public static void ComponentRenderMode(RenderTreeFrame frame, IComponentRenderMode mode, int? sequence = null)
+    public static void ComponentRenderMode(RenderTreeFrame frame, IComponentRenderMode mode)
     {
         Assert.Equal(RenderTreeFrameType.ComponentRenderMode, frame.FrameType);
         Assert.Same(mode, frame.ComponentRenderMode);
-        AssertFrame.Sequence(frame, sequence);
+        AssertFrame.Sequence(frame, 0);
     }
 
-    public static void NamedEvent(RenderTreeFrame frame, string eventType, string assignedName, int? sequence = null)
+    public static void NamedEvent(RenderTreeFrame frame, string eventType, string assignedName)
     {
         Assert.Equal(RenderTreeFrameType.NamedEvent, frame.FrameType);
         Assert.Equal(eventType, frame.NamedEventType);
         Assert.Equal(assignedName, frame.NamedEventAssignedName);
-        AssertFrame.Sequence(frame, sequence);
+        AssertFrame.Sequence(frame, 0);
     }
 }
