+++ src/Components/Web/src/Forms/Mapping/SupplyParameterFromFormValueProvider.cs
@@ -40,7 +40,7 @@ internal class SupplyParameterFromFormValueProvider : ICascadingValueSupplier
         // We also supply values for [SupplyValueFromForm]
         if (_formValueMapper is not null && parameterInfo.Attribute is SupplyParameterFromFormAttribute supplyParameterFromFormAttribute)
         {
-            return _formValueMapper.CanMap(parameterInfo.PropertyType, MappingScopeName, supplyParameterFromFormAttribute.Handler);
+            return _formValueMapper.CanMap(parameterInfo.PropertyType, MappingScopeName, supplyParameterFromFormAttribute.FormName);
         }
 
         return false;
@@ -75,7 +75,7 @@ internal class SupplyParameterFromFormValueProvider : ICascadingValueSupplier
 
         var attribute = (SupplyParameterFromFormAttribute)parameterInfo.Attribute; // Must be a valid cast because we check in CanSupplyValue
         var parameterName = attribute.Name ?? parameterInfo.PropertyName;
-        var restrictToFormName = attribute.Handler;
+        var restrictToFormName = attribute.FormName;
         Action<string, FormattableString, string?> errorHandler = string.IsNullOrEmpty(restrictToFormName) ?
             mappingContext.AddError :
             (name, message, value) => mappingContext.AddError(restrictToFormName, parameterName, message, value);
