+++ src/Components/test/testassets/Components.TestServer/RazorComponents/Components/EnableStreaming.razor
@@ -1,4 +1,4 @@
-﻿@attribute [StreamRendering(true)]
+﻿@attribute [StreamRendering]
 @ChildContent
 @code {
     [Parameter] public RenderFragment ChildContent { get; set; }
