+++ src/Components/test/testassets/Components.TestServer/RazorComponents/Pages/EnhancedNav/PageForScrollingToHash.razor
@@ -1,5 +1,5 @@
 ﻿@page "/nav/scroll-to-hash"
-@attribute [StreamRendering(true)]
+@attribute [StreamRendering]
 
 <PageTitle>Page for scrolling to hash</PageTitle>
