+++ src/Components/test/testassets/Components.TestServer/RazorComponents/Pages/EnhancedNav/PageThatRedirectsWhileStreaming.razor
@@ -1,5 +1,5 @@
 ﻿@page "/nav/do-redirection-while-streaming"
-@attribute [StreamRendering(true)]
+@attribute [StreamRendering]
 @inject NavigationManager Nav
 <h1>Please wait...</h1>
 @code {
