+++ src/Components/test/testassets/Components.TestServer/RazorComponents/Pages/EnhancedNav/PageThatThrowsWhileStreaming.razor
@@ -1,5 +1,5 @@
 ﻿@page "/nav/throw-while-streaming"
-@attribute [StreamRendering(true)]
+@attribute [StreamRendering]
 <p>Please wait...</p>
 @code {
     protected override async Task OnInitializedAsync()
