+++ src/Components/test/testassets/Components.TestServer/RazorComponents/Pages/Forms/ErrorInErrorBoundaryStreaming.razor
@@ -1,5 +1,5 @@
 ﻿@page "/forms/error-in-error-boundary-streaming"
-@attribute [StreamRendering(true)]
+@attribute [StreamRendering]
 @using Microsoft.AspNetCore.Components.Forms
 
 <h3>Error in error boundary (streaming)</h3>
