+++ src/Components/test/testassets/Components.TestServer/RazorComponents/Pages/Forms/ErrorOutsideErrorBoundaryStreaming.razor
@@ -1,5 +1,5 @@
 ﻿@page "/forms/error-outside-error-boundary-streaming"
-@attribute [StreamRendering(true)]
+@attribute [StreamRendering]
 @using Microsoft.AspNetCore.Components.Forms
 
 <h3>Error outside error boundary (streaming)</h3>
