+++ src/Components/test/testassets/Components.TestServer/RazorComponents/Pages/Forms/FormAntiforgeryAfterResponseStarted.razor
@@ -1,5 +1,5 @@
 ﻿@page "/forms/antiforgery-after-response-started"
-@attribute [StreamRendering(true)]
+@attribute [StreamRendering]
 @using Microsoft.AspNetCore.Components.Forms
 
 <h2>Default form</h2>
