+++ src/Components/test/testassets/Components.TestServer/RazorComponents/Pages/Forms/FormThatBindsGuid.razor
@@ -31,7 +31,7 @@
     string _attemptedValue = null;
     IReadOnlyList<FormattableString> _errors;
 
-    [SupplyParameterFromForm(Handler = "bind-guid")] public Guid Id { get; set; } = Guid.NewGuid();
+    [SupplyParameterFromForm(FormName = "bind-guid")] public Guid Id { get; set; } = Guid.NewGuid();
 
     [CascadingParameter] public FormMappingContext Context { get; set; }
