+++ src/Components/test/testassets/Components.TestServer/RazorComponents/Pages/Forms/NamedFormBoundParameter.razor
@@ -16,5 +16,5 @@
 @code{
     bool _submitted = false;
 
-    [SupplyParameterFromForm(Handler = "named-form-handler")] public string Parameter { get; set; } = "";
+    [SupplyParameterFromForm(FormName = "named-form-handler")] public string Parameter { get; set; } = "";
 }
