+++ src/Components/test/testassets/Components.TestServer/RazorComponents/Pages/Forms/NamedFormBoundPrimitiveParameterValidatorIntegration.razor
@@ -20,7 +20,7 @@
 @code {
     bool _submitted = false;
 
-    [SupplyParameterFromForm(Handler = "named-form-handler")] public DataModel Model { get; set; }
+    [SupplyParameterFromForm(FormName = "named-form-handler")] public DataModel Model { get; set; }
 
     [CascadingParameter] public FormMappingContext Context { get; set; }
