+++ src/Components/test/testassets/Components.TestServer/RazorComponents/Pages/Forms/PostRedirectGetStreaming.razor
@@ -1,7 +1,7 @@
 ﻿@page "/forms/post-redirect-get-streaming"
 @using Microsoft.AspNetCore.Components.Forms
 @inject NavigationManager Nav
-@attribute [StreamRendering(true)]
+@attribute [StreamRendering]
 
 <h3>Post/Redirect/Get</h3>
