+++ src/Components/test/testassets/Components.TestServer/RazorComponents/Pages/Forms/StreamingRenderingForm.razor
@@ -1,5 +1,5 @@
 ﻿@page "/forms/streaming-rendering/{OperationId}"
-@attribute [StreamRendering(true)]
+@attribute [StreamRendering]
 @inject AsyncOperationService AsyncOperation
 @inject NavigationManager Navigation;
 @using Components.TestServer.Services;
