+++ src/Components/test/testassets/Components.TestServer/RazorComponents/Pages/StreamingRendering/LargeStreamRendering.razor
@@ -1,6 +1,6 @@
 ﻿@page "/large-streaming"
 
-@attribute [StreamRendering(true)]
+@attribute [StreamRendering]
 
 <h1>Streaming Rendering</h1>
