+++ src/Components/test/testassets/Components.TestServer/RazorComponents/Pages/StreamingRendering/StreamingRendering.razor
@@ -1,6 +1,6 @@
 ﻿@page "/streaming"
 
-@attribute [StreamRendering(true)]
+@attribute [StreamRendering]
 
 <h1>Streaming Rendering</h1>
