+++ src/Components/test/testassets/Components.TestServer/RazorComponents/Pages/StreamingRendering/StreamingWithInteractivity.razor
@@ -1,6 +1,6 @@
 ﻿@page "/streaming-interactivity"
 
-@attribute [StreamRendering(true)]
+@attribute [StreamRendering]
 
 <h1>Streaming Rendering with Interactivity</h1>
